+++ blockfrost-client/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Version [next](https://github.com/blockfrost/blockfrost-haskell/compare/client-0.7.1.1...master) (2023-mm-dd)
+
+* Allow servant `0.20`
+
 # Version [0.7.1.1](https://github.com/blockfrost/blockfrost-haskell/compare/v0.7.1.0...client-0.7.1.1) (2023-01-10)
 
 * Allow blockfrost-api `0.8`